refactor(domain): align models, repositories, and use cases with backend audit
- Add ChatMessage, AddressUpdateParam, CategoryFilterParam, FavoritesPagedParam,
  FoodBulkStatusParam, MediaDownloadParam, NotificationActionParam,
  SearchRecentParam, SearchSuggestParam, VillageAddParam models
- Add CafeListUseCase, DeleteCartItemUseCase, UpdateAddressUseCase
- Remove deprecated FoodDetailRepository, GetFoodUseCase, CreateNotificationUseCase
- Update all repository interfaces with new method signatures
- Add isFavorite field to MenuSummary
- Update Stats and AppConfig domain models

+++ domain/src/main/java/com/mtv/app/shopme/domain/model/AppConfig.kt
@@ -9,8 +9,8 @@
 package com.mtv.app.shopme.domain.model
 
 data class AppConfig(
-    val minVersion: Int,
-    val latestVersion: Int,
+    val minVersion: Long,
+    val latestVersion: Long,
     val forceUpdate: Boolean,
     val maintenanceMode: Boolean,
     val maintenanceMessage: String?
